added result which seems nice to me
This result is based on my intuition and will be improved later on. Not
a lot of great math is involved, literally all the numbers are my chosen
and I am fully responsible of all the names and numbers. These are not
taken from any website just me and copilot typing it fast for me.